# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Homework 21",
        "version": "1.0.0"
    },
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "projects"
                ],
                "summary": "List projects incl. labels",
                "description": "List projects incl. labels",
                "operationId": "listProjects",
                "responses": {
                    "200": {
                        "description": "Response Json array",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "ineger",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "project1"
                                        },
                                        "user_id": {
                                            "type": "ineger",
                                            "example": "1"
                                        },
                                        "users": {
                                            "type": "JSON",
                                            "example": "[1, 2]"
                                        },
                                        "labels": {
                                            "type": "JSON",
                                            "example": "['Label1','Label2']"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "projects"
                ],
                "summary": "Add projects",
                "description": "Add project",
                "operationId": "addProject",
                "requestBody": {
                    "description": "param project",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "project1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/projects/{id}": {
            "post": {
                "tags": [
                    "projects"
                ],
                "summary": "Link project to users",
                "description": "Link project to users",
                "operationId": "syncProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Pass users id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": "[1, 2, 3]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/projects/{project}": {
            "delete": {
                "tags": [
                    "projects"
                ],
                "summary": "Delete project",
                "description": "Delete project",
                "operationId": "deleteProject",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "description": "ID of project",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    }
}